+++ python/mujinplanningclient/binpickingplanningclient.py
@@ -641,18 +641,6 @@ class BinpickingPlanningClient(realtimerobotplanningclient.RealtimeRobotPlanning
             taskparameters['ioVariableName'] = ioVariableName
         taskparameters.update(kwargs)
         return self.ExecuteCommand(taskparameters, fireandforget=fireandforget)
-    
-    def GetPackItemPoseInWorld(self, timeout=10, fireandforget=False, **kwargs):
-        """
-        Args:
-            timeout (float, optional): Time in seconds after which the command is assumed to have failed. (Default: 10)
-            fireandforget (bool, optional): If True, does not wait for the command to finish and returns immediately. The command remains queued on the server. (Default: False)
-        """
-        taskparameters = {
-            'command': 'GetPackItemPoseInWorld',
-        }
-        taskparameters.update(kwargs)
-        return self.ExecuteCommand(taskparameters, timeout=timeout, fireandforget=fireandforget)
 
     def ManuallyPlacePackItem(self, packFormationComputationResult=None, inputPartIndex=None, placeLocationNames=None, placedTargetPrefix=None, dynamicGoalsGeneratorParameters=None, orderNumber=None, numLeftToPick=None, timeout=10, fireandforget=False):
         """Places an item according to the pack formation assuming the item is placed manually and updates robotbridge state
@@ -685,3 +673,23 @@ class BinpickingPlanningClient(realtimerobotplanningclient.RealtimeRobotPlanning
         if dynamicGoalsGeneratorParameters is not None:
             taskparameters['dynamicGoalsGeneratorParameters'] = dynamicGoalsGeneratorParameters
         return self.ExecuteCommand(taskparameters, timeout=timeout, fireandforget=fireandforget)
+
+    def ValidatePackFormation(
+        self,
+        packFormation=None,
+        validationSettings=None,
+        **kwargs
+    ):
+        """Validates the pack formation.
+
+        Args:
+            packFormation (packFormationComputationResultType): A result of a pack formation computation.
+            validationSettings (dict): Parameters for the validation.
+        """
+        taskparameters = {
+            'command': 'ValidatePackFormation',
+            'packFormation': packFormation,
+            'validationSettings': validationSettings,
+        }
+        taskparameters.update(kwargs)
+        return self.ExecuteCommand(taskparameters)